fix: preserve membership database boundaries and readiness fixtures
